Network.DNS.IO

Receiving from socket

receive :: Socket -> IO DNSMessage

Receiving DNS data from Socket and parse it.

receiveVC :: Socket -> IO DNSMessage

Receive and parse a single virtual-circuit (TCP) query or response. It is up to the caller to implement any desired timeout.

Sending to socket

send :: Socket -> ByteString -> IO ()

Sending composed query or response to Socket.

sendVC :: Socket -> ByteString -> IO ()

Sending composed query or response to a single virtual-circuit (TCP).

Creating Query

encodeQuestions

Arguments

:: Identifier

-> [Question]

-> [ResourceRecord]
Additional RRs for EDNS.

-> Bool
Authentication

-> ByteString

Creating query.

composeQuery :: Identifier -> [Question] -> ByteString

Deprecated: Use encodeQuestions instead

Composing query without EDNS0.

composeQueryAD :: Identifier -> [Question] -> ByteString

Deprecated: Use encodeQuestions instead

Composing query with authentic data flag set without EDNS0.

Creating Response

responseA :: Identifier -> Question -> [IPv4] -> DNSMessage

Composing a response from IPv4 addresses

responseAAAA :: Identifier -> Question -> [IPv6] -> DNSMessage

Composing a response from IPv6 addresses