Network.DNS.Resolver

Description

Resolver related data types.

Configuration for resolver

data ResolvConf

Type for resolver configuration. Use defaultResolvConf to create a new value.

An example to use Google's public DNS cache instead of resolv.conf:

>>> let conf = defaultResolvConf { resolvInfo = RCHostName "8.8.8.8" }

An example to use multiple Google's public DNS cache concurrently:

>>> let conf = defaultResolvConf { resolvInfo = RCHostNames ["8.8.8.8","8.8.4.4"], resolvConcurrent = True }

An example to disable EDNS0:

>>> let conf = defaultResolvConf { resolvEDNS = [] }

An example to enable EDNS0 with a 1,280-bytes buffer:

>>> let conf = defaultResolvConf { resolvEDNS = [fromEDNS0 defaultEDNS0 { udpSize = 1280 }] }

An example to enable cache:

>>> let conf = defaultResolvConf { resolvCache = Just defaultCacheConf }

defaultResolvConf :: ResolvConf

Return a default ResolvConf:

resolvInfo is RCFilePath "/etc/resolv.conf".

resolvTimeout is 3,000,000 micro seconds.

resolvRetry is 3.

resolvEDNS is EDNS0 with a 4,096-bytes buffer.

resolvConcurrent is False.

resolvCache is Nothing.

Accessors

resolvInfo :: ResolvConf -> FileOrNumericHost

Server information.

resolvTimeout :: ResolvConf -> Int

Timeout in micro seconds.

resolvRetry :: ResolvConf -> Int

The number of retries including the first try.

resolvEDNS :: ResolvConf -> [ResourceRecord]

Additional resource records to specify EDNS.

resolvConcurrent :: ResolvConf -> Bool

Concurrent queries if multiple DNS servers are specified.

resolvCache :: ResolvConf -> Maybe CacheConf

Cache configuration.

Specifying DNS servers

data FileOrNumericHost

The type to specify a cache server.

Constructors

RCFilePath FilePath
A path for "resolv.conf" where one or more IP addresses of DNS servers should be found on Unix. Default DNS servers are automatically detected on Windows regardless of the value of the file name.

RCHostName HostName
A numeric IP address. Warning: host names are invalid.

RCHostNames [HostName]
Numeric IP addresses. Warning: host names are invalid.

RCHostPort HostName PortNumber
A numeric IP address and port number. Warning: host names are invalid.

Configuring cache

data CacheConf

Cache configuration for responses.

defaultCacheConf :: CacheConf

Default cache configuration.

>>> defaultCacheConf
CacheConf {maximumTTL = 300, pruningDelay = 10}

maximumTTL :: CacheConf -> TTL

If RR's TTL is higher than this value, this value is used instead.

pruningDelay :: CacheConf -> Int

Cache pruning interval in seconds.

Intermediate data type for resolver

data ResolvSeed

Intermediate abstract data type for resolvers. IP address information of DNS servers is generated according to resolvInfo internally. This value can be safely reused for withResolver.

The naming is confusing for historical reasons.

makeResolvSeed :: ResolvConf -> IO ResolvSeed

Make a ResolvSeed from a ResolvConf.

Examples:

>>> rs <- makeResolvSeed defaultResolvConf

Type and function for resolver

data Resolver

Abstract data type of DNS Resolver. This includes newly seeded identifier generators for all specified DNS servers and a cache database.

withResolver :: ResolvSeed -> (Resolver -> IO a) -> IO a

Giving a thread-safe Resolver to the function of the second argument.

withResolvers :: [ResolvSeed] -> ([Resolver] -> IO a) -> IO a

Deprecated: Use withResolver with resolvConcurrent set to True

Giving thread-safe Resolvers to the function of the second argument. For each Resolver, multiple lookups must be done sequentially. Resolvers can be used concurrently.